Shuffling monstrosity
