+++ include/a/prelude.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+#include "a/s1-lexing/lexer.hpp"
+#include "a/s1-lexing/token_matcher.hpp"
+#include "a/s1-lexing/token.hpp"
+
+#include "a/s2-parsing/ast.hpp"
+#include "a/s2-parsing/parser.hpp"
+#include "a/s2-parsing/token2ast.hpp"
+
+#include "a/s3-verification/checker.hpp"
+
+#include "a/s4-conversion/converter.hpp"
+
+#include "a/s5-codegen/emitter.hpp"
+
+#include "a/error.hpp"
