+++ include/a/s1-lexing/lexer.hpp
@@ -4,7 +4,7 @@
 #include <vector>
 #include <memory>
 
-#include "token.hpp"
+#include "token_matcher.hpp"
 
 
 /**
