+++ include/a/s3-verification/checker.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "ast.hpp"
+#include "a/s2-parsing/ast.hpp"
 
 
 class Checker {
