+++ include/a/s4-conversion/converter.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "ast.hpp"
+#include "a/s2-parsing/ast.hpp"
 
 
 class Converter {
