+++ src/error.cpp
@@ -1,4 +1,4 @@
-#include "a.hpp"
+#include "a/prelude.hpp"
 
 
 const char* CompilerError::what() const throw () {
