+++ src/meson.build
@@ -1,17 +1,19 @@
 a_sources = [
-    'a.cpp',
+    's1-lexing/lexer.cpp',
+    's1-lexing/token_matcher.cpp',
+    's1-lexing/token.cpp',
 
-    'token.cpp',
-    'lexer.cpp',
+    's2-parsing/ast.cpp',
+    's2-parsing/parser.cpp',
+    's2-parsing/token2ast.cpp',
 
-    'ast.cpp',
-    'parser.cpp',
+    's3-verification/checker.cpp',
 
-    'convert.cpp',
+    's4-conversion/converter.cpp',
 
-    'check.cpp',
+    's5-codegen/emitter.cpp',
 
-    'emit.cpp',
+    'error.cpp',
 ]
 
 a_lib = shared_library(
