+++ src/s1-lexing/lexer.cpp
@@ -1,6 +1,6 @@
 #include <sstream>
 
-#include "a.hpp"
+#include "a/prelude.hpp"
 
 
 Lexer::Lexer() {
