+++ src/s2-parsing/ast.cpp
@@ -1,4 +1,4 @@
-#include "a.hpp"
+#include "a/prelude.hpp"
 
 
 AstExpression::AstExpression(AstKind kind) :
