+++ test/prelude.hpp
@@ -6,4 +6,4 @@
 
 #include "doctest.h"
 
-#include "a.hpp"
+#include "a/prelude.hpp"
